pyproject.toml
[project]
name = "utv-legal"
version = "0.1.0"
description = ""
authors = [
    { name = "IACarvalho", email = "israelcarvalhodesenvolvedor@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.13, <4.0.0"
dependencies = ["fastapi[standard] (>=0.116.1,<0.117.0)", "sqlalchemy (>=2.0.41,<3.0.0)", "pydantic-settings (>=2.10.1,<3.0.0)", "alembic (>=1.16.4,<2.0.0)", "asyncpg (>=0.30.0,<0.31.0)", "greenlet (>=3.2.3,<4.0.0)", "psycopg2-binary (>=2.9.10,<3.0.0)", "python-jose[cryptography] (>=3.5.0,<4.0.0)", "python-multipart (>=0.0.20,<0.0.21)", "bcrypt (>=3.2.0,<4.0.0)", "passlib (>=1.7.4,<2.0.0)"]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.4"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
taskipy = "^1.14.1"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
dev = 'fastapi dev utv_legal/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=utv_legal -v'
post_test = 'coverage html'
